Make telemetry use platform only.
