+++ src/notebooks/telemetry/kernelTelemetry.ts
@@ -1,13 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { Resource } from '../platform/common/types';
-import { StopWatch } from '../platform/common/utils/stopWatch';
-import { EnvironmentType } from '../platform/pythonEnvironments/info';
-import { KernelConnectionMetadata } from '../kernels/types';
-import { Telemetry } from '../platform/common/constants';
-import { sendKernelTelemetryEvent, trackKernelResourceInformation } from './telemetry';
-import { ResourceSet } from '../platform/vscode-path/map';
+import { Resource } from '../../platform/common/types';
+import { StopWatch } from '../../platform/common/utils/stopWatch';
+import { EnvironmentType } from '../../platform/pythonEnvironments/info';
+import { KernelConnectionMetadata } from '../../kernels/types';
+import { Telemetry } from '../../platform/common/constants';
+import { sendKernelTelemetryEvent, trackKernelResourceInformation } from '../../telemetry/telemetry';
+import { ResourceSet } from '../../platform/vscode-path/map';
 
 export function sendKernelListTelemetry(
     resource: Resource,
