+++ src/platform/common/importTracker.node.ts
@@ -4,20 +4,20 @@
 
 import type * as nbformat from '@jupyterlab/nbformat';
 import { inject, injectable } from 'inversify';
-import * as path from '../platform/vscode-path/path';
+import * as path from '../vscode-path/path';
 import { NotebookCellExecutionStateChangeEvent, NotebookCellKind, NotebookDocument, TextDocument } from 'vscode';
-import { captureTelemetry, sendTelemetryEvent } from '.';
-import { splitMultilineString } from '../webviews/webview-side/common';
-import { IExtensionSingleActivationService } from '../platform/activation/types';
-import { IDocumentManager, IVSCodeNotebook } from '../platform/common/application/types';
-import { isCI, isTestExecution, PYTHON_LANGUAGE } from '../platform/common/constants';
-import '../platform/common/extensions';
-import { disposeAllDisposables } from '../platform/common/helpers';
-import { IDisposable, IDisposableRegistry } from '../platform/common/types';
-import { noop } from '../platform/common/utils/misc';
-import { EventName } from './constants';
-import { getTelemetrySafeHashedString } from './helpers';
-import { getAssociatedJupyterNotebook, isJupyterNotebook } from '../platform/common/utils';
+import { captureTelemetry, sendTelemetryEvent } from '../../telemetry';
+import { splitMultilineString } from '../../webviews/webview-side/common';
+import { IExtensionSingleActivationService } from '../activation/types';
+import { IDocumentManager, IVSCodeNotebook } from './application/types';
+import { isCI, isTestExecution, PYTHON_LANGUAGE } from './constants';
+import './extensions';
+import { disposeAllDisposables } from './helpers';
+import { IDisposable, IDisposableRegistry } from './types';
+import { noop } from './utils/misc';
+import { EventName } from '../../telemetry/constants';
+import { getTelemetrySafeHashedString } from '../../telemetry/helpers';
+import { getAssociatedJupyterNotebook, isJupyterNotebook } from './utils';
 
 /*
 Python has a fairly rich import statement. Originally the matching regexp was kept simple for
