+++ src/telemetry/serviceRegistry.node.ts
@@ -7,7 +7,6 @@ import { IExtensionSingleActivationService, IExtensionSyncActivationService } fr
 import { IServiceManager } from '../platform/ioc/types';
 import { InterpreterCountTracker } from './interpreterCountTracker';
 import { InterpreterPackages } from './interpreterPackages.node';
-import { InterpreterPackageTracker } from './interpreterPackageTracker';
 import { WorkspaceInterpreterTracker } from './workspaceInterpreterTracker';
 import { IInterpreterPackages } from './types';
 
@@ -16,10 +15,6 @@ export function registerTypes(serviceManager: IServiceManager) {
     serviceManager.addSingleton<IExtensionSyncActivationService>(
         IExtensionSyncActivationService,
         WorkspaceInterpreterTracker
-    );
-    serviceManager.addSingleton<IExtensionSyncActivationService>(
-        IExtensionSyncActivationService,
-        InterpreterPackageTracker
     );
     serviceManager.addSingleton<IExtensionSingleActivationService>(
         IExtensionSingleActivationService,
